2008-09-01  Esteban Sanchez  <estebans@artica.es>
        * extensions/update_manager.php,
        extensions/update_manager/sql/update_manager.sql: Fixed update server 
        path typo error.

        * extensions/update_manager/load_updatemanager.php: Force values to be
        integer.



git-svn-id: https://svn.code.sf.net/p/pandora/code/trunk@1062 c3f86ba8-e40f-0410-aaad-9ba5e7f4b01f

+++ pandora_console/extensions/update_manager/load_updatemanager.php
@@ -17,10 +17,10 @@ function get_user_key () {
 	/* We only want to know this for statistics records.
 	   Feel free to disable if you want. We don't want to hide anything.
 	 */
-	$n = get_db_value ('COUNT(`id_agente`)', 'tagente', 'disabled', 0);
-	$m = get_db_value ('COUNT(`id_agente_modulo`)', 'tagente_modulo',
-			'disabled', 0);
-	$user_key = array ('A' => $n, 'M' => $m);
+	$n = (int) get_db_value ('COUNT(`id_agente`)', 'tagente', 'disabled', 0);
+	$m = (int) get_db_value ('COUNT(`id_agente_modulo`)', 'tagente_modulo',
+				'disabled', 0);
+	$user_key = array ('A' => )$n, 'M' => $m);
 	
 	return json_encode ($user_key);
 }
